# documentation/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Search API
paths:
...
  /v2.0/search/addresses/postcode/{postcode}:
    get:
      summary: Get addresses by postcode
      description: Get addresses by postcode
      parameters:
      - name: postcode
        in: path
        type: string
        required: true
        description: Postcode to search on
      - name: base64
        in: query
        type: string
        required: false
        description: if "true" postcode will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            type: array
            items:
              $ref: '#/definitions/mapped_address'
        400:
          description: Invalid postcode format
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Invalid input
              error_code:
                type: number
                example: 400
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No addresses found for search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
  /v2.0/search/addresses/uprn/{uprn}:
    get:
      summary: Get addresses by uprn
      description: Get addresses by uprn
      parameters:
      - name: uprn
        in: path
        type: string
        required: true
        description: UPRN to search on
      - name: base64
        in: query
        type: string
        required: false
        description: if "true" uprn will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            type: array
            items:
              $ref: '#/definitions/mapped_address'
        400:
          description: Invalid postcode format
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Invalid input
              error_code:
                type: number
                example: 400
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No addresses found for search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
  /v2.0/search/addresses/usrn/{usrn}:
    get:
      summary: Get addresses by usrn
      description: Get addresses by usrn
      parameters:
      - name: usrn
        in: path
        type: string
        required: true
        description: USRN to search on
      - name: base64
        in: query
        type: string
        required: false
        description: if "true" usrn will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            type: array
            items:
              $ref: '#/definitions/mapped_address'
        400:
          description: Invalid postcode format
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Invalid input
              error_code:
                type: number
                example: 400
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No addresses found for search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
  /v2.0/search/addresses/text/{text}:
    get:
      summary: Get addresses by text
      description: Get addresses by text
      parameters:
      - name: text
        in: path
        type: string
        required: true
        description: Text to search on
      - name: base64
        in: query
        type: string
        required: false
        description: if "true" text will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            type: array
            items:
              $ref: '#/definitions/mapped_address'
        400:
          description: Invalid postcode format
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Invalid input
              error_code:
                type: number
                example: 400
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No addresses found for search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
  /v2.0/search/addresses/title/{title}:
    get:
      summary: Get index map by title
      description: Get index map by title
      parameters:
        - name: title
          in: path
          type: string
          required: true
          description: title to search on
        - name: base64
          in: query
          type: string
          required: false
          description: if "true" title will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            $ref: '#/definitions/FeatureCollection'
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No index map found for title search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
...
  /search/addresses/postcode/{postcode}:
    get:
      summary: Get addresses by postcode
      description: Get addresses by postcode
      parameters:
      - name: postcode
        in: path
        type: string
        required: true
        description: Postcode to search on
      - name: base64
        in: query
        type: string
        required: false
        description: if "true" postcode will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            type: array
            items:
              $ref: '#/definitions/mapped_address'
        400:
          description: Invalid postcode format
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Invalid input
              error_code:
                type: number
                example: 400
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No addresses found for search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
  /search/addresses/uprn/{uprn}:
    get:
      summary: Get addresses by uprn
      description: Get addresses by uprn
      parameters:
      - name: uprn
        in: path
        type: string
        required: true
        description: UPRN to search on
      - name: base64
        in: query
        type: string
        required: false
        description: if "true" uprn will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            type: array
            items:
              $ref: '#/definitions/mapped_address'
        400:
          description: Invalid postcode format
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Invalid input
              error_code:
                type: number
                example: 400
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No addresses found for search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
  /search/addresses/usrn/{usrn}:
    get:
      summary: Get addresses by usrn
      description: Get addresses by usrn
      parameters:
      - name: usrn
        in: path
        type: string
        required: true
        description: USRN to search on
      - name: base64
        in: query
        type: string
        required: false
        description: if "true" usrn will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            type: array
            items:
              $ref: '#/definitions/mapped_address'
        400:
          description: Invalid postcode format
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Invalid input
              error_code:
                type: number
                example: 400
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No addresses found for search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
  /search/addresses/text/{text}:
    get:
      summary: Get addresses by text
      description: Get addresses by text
      parameters:
      - name: text
        in: path
        type: string
        required: true
        description: Text to search on
      - name: base64
        in: query
        type: string
        required: false
        description: if "true" text will be safely decoded in base64
      responses:
        200:
          description: Search successful
          schema:
            type: array
            items:
              $ref: '#/definitions/mapped_address'
        400:
          description: Invalid postcode format
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Invalid input
              error_code:
                type: number
                example: 400
        404:
          description: No results found
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: No addresses found for search
              error_code:
                type: number
                example: 404
        500:
          description: Server error
          schema:
            type: object
            properties:
              error_message:
                type: string
                example: Something went wrong
              error_code:
                type: number
                example: 500
security:
  - JWTAuth: []
securityDefinitions:
  JWTAuth:
    type: apiKey
    in: header
    name: Authorizaton
definitions:
...
  FeatureCollection:
    description: GeoJSON FeatureCollection
    type: object
    properties:
      type:
        type: string
        enum:
        - 'FeatureCollection'
      features:
        type: array
        items:
          $ref: '#/definitions/Feature'
  Feature:
    description: GeoJSON Feature
    type: object
    properties:
      geometry:
        $ref: '#/definitions/Polygon'
      properties:
        type: object
        properties:
          id:
            type: integer
      type:
        type: string
        enum:
        - 'Feature'
  Point2D:
    type: array
    description: 2D point
    minItems: 2
    maxItems: 2
    items:
      type: number
  Polygon:
    type: object
    description: GeoJSON polygon
    properties:
      coordinates:
        type: array
        items:
          type: array
          items:
            $ref: '#/definitions/Point2D'
      type:
        type: string
        enum:
        - 'Polygon'
  mapped_address:
    description: GeoJSON Feature
    type: object
    properties:
      address_type:
        type: string
        example: property
      address:
        type: string
        example: 1 Test street, Town, County, TT1 1TT
      geometry:
        $ref: '#/definitions/FeatureCollection'
      postcode:
        type: string
        example: TT1 1TT
      uprn:
        type: string
        example: 123ABC
      line_1:
        type: string
      line_2:
        type: string
      line_3:
        type: string
      line_4:
        type: string
      line_5:
        type: string
      line_6:
        type: string
      street:
        type: string
      town:
        type: string